requirements.txt
streamlit
langchain
selenium
beautifulsoup4
requests
lxml
html5lib
python-dotenv
langchain-google-genai
